refactor: organize constants
