+++ packages/client/src/constants/configs/leaderboard/index.ts
